UI Overhaul 1/
Moving Activity contents to Fragments
Unified XML resources for all Activities and Fragments (Base)
