+++ app/src/main/java/com/example/android/miwok/PhrasesFragment.kt
@@ -1,0 +1,71 @@
+package com.example.android.miwok
+
+
+import android.media.MediaPlayer
+import android.os.Bundle
+import android.support.v4.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.AdapterView
+import android.widget.ListView
+import kotlinx.android.synthetic.main.fragment_base.*
+
+class PhrasesFragment : Fragment() {
+    private lateinit var mMediaPlayer : MediaPlayer
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        // Inflate the layout for this fragment
+        val view = inflater.inflate(R.layout.fragment_base, container,false)
+
+        val words = ArrayList<Word>()
+        words.add(Word("Where are you going?", "minto wuksus", R.raw.phrase_where_are_you_going))
+        words.add(Word("What is your name?", "tinnә oyaase'nә", R.raw.phrase_what_is_your_name))
+        words.add(Word("My name is...", "oyaaset...", R.raw.phrase_my_name_is))
+        words.add(Word("How are you feeling?", "michәksәs?", R.raw.phrase_how_are_you_feeling))
+        words.add(Word("I’m feeling good.", "kuchi achit", R.raw.phrase_im_feeling_good))
+        words.add(Word("Are you coming?", "әәnәs'aa?", R.raw.phrase_are_you_coming))
+        words.add(Word("Yes, I’m coming.", "hәә’ әәnәm", R.raw.phrase_yes_im_coming))
+        words.add(Word("I’m coming.", "әәnәm", R.raw.phrase_im_coming))
+        words.add(Word("Let’s go.", "yoowutis", R.raw.phrase_lets_go))
+        words.add(Word("Come here.", "әnni'nem", R.raw.phrase_come_here))
+
+        // Create an {@link WordAdapter}, whose data source is a list of {@link Word}s. The
+        // adapter knows how to create list items for each item in the list.
+        val adapter = WordAdapter(activity!!.applicationContext, words)
+
+        // Find the {@link ListView} object in the view hierarchy of the {@link Activity}.
+        // There should be a {@link ListView} with the view ID called list, which is declared in the
+        // word_list.xml layout file.
+        val listView = view.findViewById<ListView>(R.id.list)
+
+        // Make the {@link ListView} use the {@link WordAdapter} we created above, so that the
+        // {@link ListView} will display list items for each {@link Word} in the list.
+        listView.adapter = adapter
+
+        listView.onItemClickListener = AdapterView.OnItemClickListener { _, _, position, _ ->
+            // Get the {@link Word} object at the given position the user clicked on
+            // Create and setup the {@link MediaPlayer} for the audio resource associated
+            // with the current word
+
+            mMediaPlayer = MediaPlayer.create(activity, words[position].getAudioResourceId())
+
+            // Start the audio file
+            mMediaPlayer.start()
+            mMediaPlayer.setOnCompletionListener {
+                mMediaPlayer.release()
+            }
+
+        }
+
+        return view
+    }
+
+    override fun onStop() {
+        super.onStop()
+        if(this::mMediaPlayer.isInitialized)
+            mMediaPlayer.release()
+    }
+}
+
